backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
google-generativeai>=0.8.0
google-cloud-speech>=2.21.0
google-auth>=2.23.0
requests>=2.31.0
sqlalchemy==2.0.25
aiosqlite==0.19.0
pydantic==2.5.3

